backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
sqlalchemy==2.0.40
asyncpg==0.30.0
alembic==1.15.2
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
